requirements.txt
pyrogram
TgCrypto
requests
urllib3
jikanpy
ffmpeg-python